% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pondr.R
\name{extract_pondr}
\alias{extract_pondr}
\title{Extract data from text files of protein disorder predictors from PONDR.}
\usage{
extract_pondr(directory, output = NULL)
}
\arguments{
\item{directory}{A character string describing the path to a Directory where
all the text files are located. Note, the function will attempt to read every
file in the directory so it should contain only the files you wish to be read.}

\item{output}{A character string describing the path to write the output
matrix. If `NULL`, an output file will not be written.}
}
\value{
A matrix containing, for each protein, the Predicted residues ("resid"), Number
Disordered Regions ("dis_rgns"), Number residues disordered ("n_dis"),
Longest Disordered Region ("lg_rgn"), Overall percent disordered ("pct"),
and Average Prediction Score ("avg") for each of the three predictors.
}
\description{
This function extracts numerical data from text porduced by the VLXT, VL3,
and VSL2 disorder predictors from the Predictor of Natural Disordered Regions
 (PONDR) website <http://www.pondr.com/>. For each protien sequence, the data
 from the Raw Output of all three predictors were pasted into a .txt file.
}
\examples{
extract_pondr("data/pondr_text/")
}
