% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pondrFIT.R
\name{extract_pondrFIT}
\alias{extract_pondrFIT}
\title{Extract data from URLs produced by PONDR-FIT meta-predictor}
\usage{
extract_pondrFIT(path, save_raw = NULL, output = NULL)
}
\arguments{
\item{path}{A character string desctibing the path to the csv file containing URLs to be read.}

\item{save_raw}{A character string describing the path to a *Directory* where intermediate files for each protein shoul be saved. If `NULL`, the intermediate files will not be written to csv.}

\item{output}{A character string describing the path to write the output matrix to. If `NULL`, an output file will not be written.}
}
\value{
a martix containing the average disorder scores, % disorder scores, and length (in addition to the UniProt ID and URL)
}
\description{
This function extracts the relevant data from the temporaty URL produced by
analyzing a protein sequence in the PONDR-FIT protein disorder meta-predictor
(http://original.disprot.org/pondr-fit.php). URLs were collected and put in a
csv file with the UniProt ID in the first column and URL in the second.
}
\examples{
extract_pondrFIT("./Data/pondrfit-url.csv")
}
