% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_string.R
\name{extract_string}
\alias{extract_string}
\title{Extract data on the quantity of protein interactions given by STRING.}
\usage{
extract_string(direct, output = NULL)
}
\arguments{
\item{direct}{A character string describing the path to a Directory where
all the text files are located. Note, the function will attempt to read every
file in the directory so it should contain only the files you wish to be read.}

\item{output}{A character string describing the path to write the output
matrix. If `NULL`, an output file will not be written.}
}
\value{
A matrix containing the network stats at each of the three confidence
levels for every protein.
}
\description{
Extract data on the quantity of protein interactions given by STRING
(https://string-db.org/) with the minimum number of interactions at
medium (0.4), high (0.7), and highest (0.9) confidence.
}
\examples{
extract_string("inst/extdata/string/")
}
