% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pondr_data}
\alias{pondr_data}
\title{Data on disordered regions in proteins from several predictors on PONDR.}
\format{
A matrix with 10 rows and 18 variables
}
\usage{
data(pondr_data)
}
\description{
An example dataset produced by the extract_pondr() function containing the
attributes on disorder in proteins calculated by several prodein disorder
predictors on PONDR /url{http://www.pondr.com/}.
}
\details{
\itemize{
  \item VLXT, VL3, VSL2. protein disorder predictors on PONDR
  \item resid. Predicted residues (total number)
  \item dis_rgns. Number Disordered Regions
  \item n_dis. Number residues disordered
  \item lg_rgn. Longest Disordered Region
  \item pct. Overall percent disordered
  \item avg. Average Prediction Score
}
}
\keyword{datasets}
