% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pondrFIT_data}
\alias{pondrFIT_data}
\title{Data on disordered regions in proteins from PONDR-FIT meta-predictor.}
\format{
A data frame with 10 rows and 5 variables
}
\usage{
data(pondrFIT_data)
}
\description{
An example dataset produced by the extract_pondrFIT() function containing the
attributes on disorder in proteins calculated by the PONDR-FIT disorder
meta-predictors on DisPort /url{http://original.disprot.org/pondr-fit.php}.
}
\details{
\itemize{
  \item UniprotID. Unique identifier of the protein in Uniprot database /url{https://www.uniprot.org/}
  \item url. temporaty URL of the raw data produced by analyzing a protein sequence with PONDR-FIT
  \item meanDisorder. The average (mean) of the per-residue disorder scores
  \item percentDisorder. The proportion of the per-residue scores that are greater than 0.5
  \item lenngth. A count of the number of residues in the protein sequence
}
}
\keyword{datasets}
