% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string_data}
\alias{string_data}
\title{Data on protein interaction networks.}
\format{
A matrix with 10 rows and 18 variables
}
\usage{
data(string_data)
}
\description{
An example dataset produced by the extract_string() function containing the
attributes on protein interactions calculated by several confidence levels
on STRING /url{https://string-db.org/}.
}
\details{
\itemize{
  \item 0.4, 0.7, 0.9. confidence levels
  \item nodes. Number of nodes (proteins) in network
  \item edges. Number of edges (interactions) in network
  \item node_deg. Average node degree
  \item lc_coef. Avgerage local clustering coeffficient
  \item edges_Exp. Expected number of edges
  \item p_PPI. PPI enrichment p-value
}
}
\keyword{datasets}
